NexusEngine-FINAL/NexusEngine/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
psutil>=5.9.0
